feat: hash operator for Lair
To realize a hash operator for Lair, the `Hasher` trait had to be
extended in order to provide enough data for Air-related impls.

`Op::Hash` needs access to a `is_real` for the block it belongs to.
`is_real` can be defined as the sum of all `local.sel[ident]` for
every `ident` in `Ctrl::Return(ident, _)` that block (transitively)
contains.

In order to accomplish the above, `return_idents: List<usize>` was
added as a new attribute of `Block`, computed during compilation.