feat: Change default nums to u64, add n suffix
Also removes the "1/2" convenience syntax parsing. Previously, "1/2"
would return the same result as "(/ 1 2)". This behavior does not really
make as much sense with u64s as it does with field elements, so it's
removed for now. The parsing with `n` suffix for field elements does not
easily support the `1n/2n` syntax since it doesn't parse the second
suffix. This can be re-added later if desired.