Change default numeric type to u64 (#181)
* feat: Change default nums to u64, add n suffix

Also removes the "1/2" convenience syntax parsing. Previously, "1/2"
would return the same result as "(/ 1 2)". This behavior does not really
make as much sense with u64s as it does with field elements, so it's
removed for now. The parsing with `n` suffix for field elements does not
easily support the `1n/2n` syntax since it doesn't parse the second
suffix. This can be re-added later if desired.

* feat: Add u64_iszero, fix div by zero check

* fix: Fix the REPL output to match new defaults

* chore: Update loam tests for the new syntax

* chore: Remove unnecessary & in assert_is_zero

* chore: Add additional u64_iszero test case

* chore: Add u64 chips to test_widths

---------

Co-authored-by: wwared <541936+wwared@users.noreply.github.com>