feat: `Lang` and custom gadgets
A whole bunch of changes in order to enable user-made gadgets
through `Lang`.

Differently from Lurk Beta and R1CS, this patch empowers the user
with the full blown expressiveness of Lair - not just Air.

There are changes in Lair to support an extra `Chipset` type, but
most of the significant changes are in Lurk itself.