fix: memoize dag of reduced call args (#379)
Call-related meta commands now have their arguments evaluated before
constructing the final call expression. This was causing a hole in
the ZStore's DAG. To fix the hole, we need to memoize the DAG of each
reduced argument separately.

This solution is replicated for the protocol API, which could suffer
from the same kind of problem. Memoizing the DAG of the full list
wasn't enough because the memoizatin algorithm would get interrupted
from hitting the cache. We must memoize from the bottom up. Then,
`ZStore::intern_list` will take care of the rest (no manual
memoization is needed).

Extra:
* Simplify the `Repl::memoize_dag` API
* Optimize the `def` meta command to simply evaluate the expression
  instead of macro-expanding it into a `let` expression that returns
  `(current-env)`